requirements.txt
pyautogui
numpy
opencv-python
mediapipe
joblib
scikit-learn
streamlit
pyserial